+++ src/city.c
@@ -4,11 +4,11 @@
 #include <stdlib.h>
 
 struct city {
-  size_t id;
-  double x, y;
+  uint16_t id;
+  float x, y;
 };
 
-City *city_new(size_t id, double x, double y) {
+City *city_new(uint16_t id, float x, float y) {
   City *c = malloc(sizeof(City));
   c->id = id;
   c->x = x;
@@ -16,29 +16,29 @@ City *city_new(size_t id, double x, double y) {
   return c;
 }
 
-size_t city_id(City *c) { return c->id; }
+uint16_t city_id(City *c) { return c->id; }
 
-double city_x(City *c) { return c->x; }
+float city_x(City *c) { return c->x; }
 
-double city_y(City *c) { return c->y; }
+float city_y(City *c) { return c->y; }
 
 void city_free(City *c) {
   assert(c != NULL && "city must not be NULL");
   free(c);
 }
 
-double city_calculate_distance(City *c1, City *c2) {
-  double dx = c2->x - c1->x;
-  double dy = c2->y - c1->y;
+float city_calculate_distance(City *c1, City *c2) {
+  float dx = c2->x - c1->x;
+  float dy = c2->y - c1->y;
   return sqrt(pow(dx, 2) + pow(dy, 2));
 }
 
-City **city_array_new(size_t n) { return malloc(n * sizeof(City)); }
+City **city_array_new(uint16_t n) { return malloc(n * sizeof(City)); }
 
-void cities_free(City **cities, size_t qtd_cities) {
+void cities_free(City **cities, uint16_t qtd_cities) {
   assert(cities != NULL && "cities must not be NULL");
 
-  for (size_t i = 0; i < qtd_cities; i++) {
+  for (uint16_t i = 0; i < qtd_cities; i++) {
     if (!cities[i])
       continue;
     city_free(cities[i]);
